# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentic Workflow Engine",
    "description": "Agentic Workflow Automation Engine API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/executions/": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "List Executions",
        "operationId": "list_executions_api_v1_executions__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowRunRead"
                  },
                  "title": "Response List Executions Api V1 Executions  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/executions/{run_id}": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "Get Execution Run",
        "operationId": "get_execution_run_api_v1_executions__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/executions/{run_id}/logs": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "Get Execution Step Logs",
        "operationId": "get_execution_step_logs_api_v1_executions__run_id__logs_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepLogRead"
                  },
                  "title": "Response Get Execution Step Logs Api V1 Executions  Run Id  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/executions/{run_id}/cancel": {
      "post": {
        "tags": [
          "Executions"
        ],
        "summary": "Cancel Execution Run",
        "operationId": "cancel_execution_run_api_v1_executions__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credentials/": {
      "get": {
        "tags": [
          "Credentials"
        ],
        "summary": "List Credentials",
        "operationId": "list_credentials_api_v1_credentials__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CredentialRead"
                  },
                  "type": "array",
                  "title": "Response List Credentials Api V1 Credentials  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Credentials"
        ],
        "summary": "Create Credential",
        "operationId": "create_credential_api_v1_credentials__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credentials/{credential_id}": {
      "delete": {
        "tags": [
          "Credentials"
        ],
        "summary": "Delete Credential",
        "operationId": "delete_credential_api_v1_credentials__credential_id__delete",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Credential Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CredentialCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "service_type": {
            "type": "string",
            "title": "Service Type"
          },
          "raw_secret": {
            "type": "string",
            "title": "Raw Secret"
          }
        },
        "type": "object",
        "required": [
          "name",
          "service_type",
          "raw_secret"
        ],
        "title": "CredentialCreate"
      },
      "CredentialRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "service_type": {
            "type": "string",
            "title": "Service Type"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "service_type",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "CredentialRead"
      },
...
      "ExecutionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "ExecutionStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "StepLogRead": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "node_name": {
            "type": "string",
            "title": "Node Name"
          },
          "node_type": {
            "type": "string",
            "title": "Node Type"
          },
          "status": {
            "$ref": "#/components/schemas/ExecutionStatus",
            "default": "pending"
          },
          "input_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Data"
          },
          "output_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output Data"
          },
          "thought_trace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thought Trace"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "execution_time_ms": {
            "type": "number",
            "title": "Execution Time Ms",
            "default": 0.0
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "node_id",
          "node_name",
          "node_type",
          "id",
          "timestamp"
        ],
        "title": "StepLogRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowRunRead": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "status": {
            "$ref": "#/components/schemas/ExecutionStatus",
            "default": "pending"
          },
          "trigger_type": {
            "type": "string",
            "title": "Trigger Type",
            "default": "manual"
          },
          "input_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Data"
          },
          "output_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output Data"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "workflow_id",
          "id",
          "started_at"
        ],
        "title": "WorkflowRunRead"
      },
...
    }
  }
}